+++ L02/Acme-Madruga/src/main/java/converters/FloatToStringConverter.java
@@ -4,13 +4,13 @@ package converters;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.stereotype.Component;
 
-import domain.Float;
+import domain.BigDecimal;
 
 @Component
-public class FloatToStringConverter implements Converter<domain.Float, String> {
+public class FloatToStringConverter implements Converter<domain.BigDecimal, String> {
 
 	@Override
-	public String convert(final domain.Float floatt) {
+	public String convert(final domain.BigDecimal floatt) {
 		String result;
 
 		if (floatt == null)
