+++ L02/Acme-Madruga/src/main/java/repositories/FloatRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import domain.Float;
+import domain.BigDecimal;
 
 /*
  * CONTROL DE CAMBIOS ProcessionRepository.java
@@ -17,8 +17,8 @@ import domain.Float;
  */
 
 @Repository
-public interface FloatRepository extends JpaRepository<Float, Integer> {
+public interface FloatRepository extends JpaRepository<BigDecimal, Integer> {
 
 	@Query("select f from domain.Float f where f.brotherhood.id=?1")
-	Collection<domain.Float> findFloatByBrotherhood(int brotherhoodId);
+	Collection<domain.BigDecimal> findFloatByBrotherhood(int brotherhoodId);
 }
